01_Crawl4AI_Scraper/requirements.txt
httpx
playwright
beautifulsoup4
lxml
python-dotenv
aiofiles